Try to fix a type error in model.py
Start training...
Traceback (most recent call last):
  File "main.py", line 64, in <module>
    main(config)
  File "main.py", line 33, in main
    solver.train()
  File "/home/user/PycharmProjects/SpeechSplit-master/solver.py", line 161, in train
    x_f0_intrp = self.Interp(x_f0, len_org)
  File "/home/user/anaconda3/envs/vencen/lib/python3.6/site-packages/torch/nn/modules/module.py", line 547, in __call__
    result = self.forward(*input, **kwargs)
  File "/home/user/PycharmProjects/SpeechSplit-master/model.py", line 395, in forward
    idx_scaled = indices/ scales.unsqueeze(-1)
RuntimeError: expected device cuda:0 and dtype Float but got device cuda:0 and dtype Long
Aborted (core dumped)
exit status 134

+++ model.py
@@ -392,7 +392,7 @@ class InterpLnr(nn.Module):
         scales = torch.rand(batch_size*self.max_num_seg, 
                             device=device) + 0.5
         
-        idx_scaled = indices / scales.unsqueeze(-1)
+        idx_scaled = indices.float()  / scales.unsqueeze(-1)
         idx_scaled_fl = torch.floor(idx_scaled)
         lambda_ = idx_scaled - idx_scaled_fl
